Sorted hash collision detection

+++ KittenEngine/YarnBall/YarnBall.cu
@@ -136,7 +136,7 @@ namespace YarnBall {
 		cudaMemset(meta.d_lastVels, 0, sizeof(vec3) * numVerts);
 
 		cudaMalloc(&meta.d_hashTable, sizeof(int) * meta.hashTableSize);
-		cudaMalloc(&meta.d_numCols, sizeof(uint16_t) * numVerts);
+		cudaMalloc(&meta.d_numCols, sizeof(int) * numVerts);
 		cudaMalloc(&meta.d_collisions, sizeof(Collision) * numVerts * MAX_COLLISIONS_PER_SEGMENT);
 
 		vertBuffer = new Kitten::CudaComputeBuffer(sizeof(Vertex), numVerts);
